cpu-boost: Skip deactivating shadow if a boost is pending
When a new input boost is queued,
it cannot cancel the work to deactivate
the input boost if it is already running.
The runtime of that work can be relatively
long, due to the needed policy updates and
now with shadow, also the needed task
reclassification.

There are two problems resulting from this:
1. The runtime of the work to deactivate
   the input boost is too long and can
   cause an additional latency on the
   next input boost work (since it is synced).
2. A full task reclassification and also
   the resulting task migration are
   expensive tasks, so it would be less
   expensive to just skip the deactivation
   of shadow, if it is going to be reactivated
   after some ms anyways.

To fix these problems, this patch introduces
"input_boost_pending" and skips deactivating
shadow if a new input boost is waiting for
the completion of the input boost deactivation
work.

Signed-off-by: Alex Naidis <alex.naidis@linux.com>