ARM: dts: msm8994: Use the correct adc channel for msm_thermal_simple I figured out channel 115 is msm_therm, which is the same temperature thermald is monitoring.
With this channel I was able to run high synthetics load benchmark
without having the phone shutting down because of the temperature.

NOTA: for some reason, channel 117 (Which should represent pa_therm0)
always return -40 for me. Not sure if it's a bug in my kernel or it's
a common problem of msm8994. More investigation about it would be
required. pa_therm1 works fine though.